count rows as separate method

+++ src/main/java/io/github/smolcan/aggrid/jpa/adapter/response/LoadSuccessParams.java
@@ -7,13 +7,13 @@ public class LoadSuccessParams {
     // Data retrieved from the server as requested by the grid.
     private List<Map<String, Object>> rowData;
     // The last row, if known, to help Infinite Scroll.
-    private Integer rowCount;
+    private Long rowCount;
     // Any extra information for the grid to associate with this load.
     private Map<String, Object> groupLevelInfo;
     // The pivot fields in the response - if provided the grid will attempt to generate secondary columns.
     private List<String> pivotResultFields;
 
-    public LoadSuccessParams(List<Map<String, Object>> rowData, Integer rowCount, Map<String, Object> groupLevelInfo, List<String> pivotResultFields) {
+    public LoadSuccessParams(List<Map<String, Object>> rowData, Long rowCount, Map<String, Object> groupLevelInfo, List<String> pivotResultFields) {
         this.rowData = rowData;
         this.rowCount = rowCount;
         this.groupLevelInfo = groupLevelInfo;
@@ -31,11 +31,11 @@ public class LoadSuccessParams {
         this.rowData = rowData;
     }
 
-    public Integer getRowCount() {
+    public Long getRowCount() {
         return rowCount;
     }
 
-    public void setRowCount(Integer rowCount) {
+    public void setRowCount(Long rowCount) {
         this.rowCount = rowCount;
     }
